+++ sdk/include/model/object/ListedCommonPrefix.h
@@ -12,6 +12,25 @@ public:
         prefix_ = prefix;
     }
 
+    bool operator==(const ListedCommonPrefix& rhs) const {
+        return prefix_ == rhs.prefix_;
+    }
+    bool operator!=(const ListedCommonPrefix& rhs) const {
+        return !(rhs == *this);
+    }
+    bool operator<(const ListedCommonPrefix& rhs) const {
+        return prefix_ < rhs.prefix_;
+    }
+    bool operator>(const ListedCommonPrefix& rhs) const {
+        return rhs < *this;
+    }
+    bool operator<=(const ListedCommonPrefix& rhs) const {
+        return !(rhs < *this);
+    }
+    bool operator>=(const ListedCommonPrefix& rhs) const {
+        return !(*this < rhs);
+    }
+
 private:
     std::string prefix_;
 };