+++ CHANGELOG.md
@@ -1,8 +1,29 @@
 ## Release Note
 
+### 2022.10.31 Version 2.4.0
+
+- add: client DNS cache
+- add: proxy is supported
+- add: listObjectsType2 interface
+- add: bucket lifecycle related interfaces
+- add: bucket policy related interface
+- add: bucket storage type interface
+- add: bucket CORS interface
+- add: bucket mirror back interface
+- add: bucket ACL interface
+- add: object tag interface
+- add: fetch object interface
+- change: unified exception error definition
+- fix: downloadFile defect
+
 ### 2022.10.7 Version 2.2.0
 
 - add clientV2
+- add: breakpoint resume download
+- add: upload and download can display progress bar
+- add: speed limit for upload and download
+- add: client CRC verification
+- add: client disconnection verification
 
 ### 2022.5.30 Version 0.2.1